rename: kakao-map -> signup-kakao-map(#306)
